Separate Event machine and State machine concepts
