+++ app/src/main/java/io/github/gymbay/gelm/app/example/store/components/ExampleInternalReducer.kt
@@ -4,7 +4,6 @@ import io.github.gymbay.gelm.app.example.store.models.ExampleCommand
 import io.github.gymbay.gelm.app.example.store.models.ExampleEffect
 import io.github.gymbay.gelm.app.example.store.models.ExampleInternalEvent
 import io.github.gymbay.gelm.app.example.store.models.ExampleState
-import io.github.gymbay.gelm.app.example.store.models.ExampleStateMachine
 import io.github.gymbay.gelm.reducers.GelmInternalReducer
 import io.github.gymbay.gelm.reducers.Modifier
 
@@ -23,38 +22,4 @@ class ExampleInternalReducer :
             }
         }
     }
-}
-
-private typealias InternalStateModifier = Modifier<ExampleStateMachine, ExampleEffect, ExampleCommand>
-
-class ExampleInternalReducerStateMachine :
-    GelmInternalReducer<ExampleInternalEvent, ExampleStateMachine, ExampleEffect, ExampleCommand>() {
-
-    override fun Modifier<ExampleStateMachine, ExampleEffect, ExampleCommand>.processInternalEvent(
-        currentState: ExampleStateMachine,
-        internalEvent: ExampleInternalEvent
-    ) {
-        when (currentState) {
-            is ExampleStateMachine.Idle -> Unit
-            is ExampleStateMachine.Fetching -> processFetchingState(currentState, internalEvent)
-        }
-    }
-
-    private fun InternalStateModifier.processFetchingState(
-        fetchingState: ExampleStateMachine.Fetching,
-        event: ExampleInternalEvent
-    ) {
-        when (event) {
-            is ExampleInternalEvent.LoadedData -> {
-                state {
-                    ExampleStateMachine.Idle(
-                        title = fetchingState.title,
-                        editField = fetchingState.editField,
-                        items = event.list
-                    )
-                }
-            }
-        }
-    }
-
 }
